chore: move initialiseListeners to standalone file to avoid circular deps
